ticket-1577 IMA Agent moved into it's own repository
