+++ package.json
@@ -4,11 +4,11 @@
     "license": "AGPL-3.0",
     "author": "SKALE Labs and contributors",
     "scripts": {
-        "lint-check": "cd src; yarn lint-check; cd ..",
-        "lint-fix": "cd src; yarn lint-fix; cd ..",
-        "postinstall": "cd src; yarn install; cd ..",
-        "check-outdated": "yarn outdated; cd src; yarn outdated; cd ..",
-        "upgrade-to-latest": "yarn upgrade --latest; cd src; yarn upgrade --latest; cd .."
+        "lint-check": "cd src; yarn lint-check; cd ..; cd IMA; yarn lint-check; cd ..",
+        "lint-fix": "cd src; yarn lint-fix; cd ..; cd IMA; yarn lint-fix; cd ..",
+        "postinstall": "cd src; yarn install; cd ..; cd IMA; yarn install; cd ..",
+        "check-outdated": "yarn outdated; cd src; yarn outdated; cd ..; cd IMA; yarn outdated; cd ..",
+        "upgrade-to-latest": "yarn upgrade --latest; cd src; yarn upgrade --latest; cd ..; cd IMA; yarn upgrade --latest; cd .."
     },
     "dependencies": {},
     "devDependencies": {}
